Collision System now uses bvh queries for finding closest colliders to test intersect against.
Howl app now has debug stop watches for update steps.

Added deallocate all function to gen index list.

Added query function to BVH to find nearest entries.

BVH leaf construction optimised by removing intial entry insertion checks.

Fixed leaf nodes not having entries count properly set.

Fixed bug with renderer not drawing primitive amount beyond short.MaxValue even though primitives stored as ints.

+++ src/vendors/monogame/graphics/Renderer.cs
@@ -651,7 +651,7 @@ public sealed class Renderer : IRenderer
 
     public void DrawWireframeShape(in Howl.Math.Transform transform, in Howl.Graphics.RectangleShape rectangle, float thickness = 4)
     {
-        if(primitiveVertices.Count > short.MaxValue)
+        if(primitiveVertices.Count > int.MaxValue)
         {
             throw new OverflowException();
         }
